feat: refactor auction models and update MQTT client for price handling
