+++ backend/auction/mqtt_client.py
@@ -11,29 +11,31 @@ TOPIC = "auction/#"
 
 client = None  # Global client variable
 
+PRICE_CHANGE = 50
+
 
 def on_connect(client, userdata, flags, rc):
     print("Connected to MQTT Broker")
     client.subscribe(TOPIC)
 
 
-def on_message(client, userdata, msg):
+def on_message(client, userdata, msg) -> None:
     payload = msg.payload.decode('utf-8')
     topic = msg.topic
 
     if topic.startswith("auction/"):
         auction_id = topic.split('/')[1]
         try:
-            auction = Auction.objects.get(id=auction_id, is_active=True)
-            user = User.objects.get(card_id=payload)
+            auction: Auction = Auction.objects.get(
+                id=auction_id, is_active=True)
+            user: User = User.objects.get(card_id=payload)
 
-            if user.balance >= auction.item.current_price + 1:
-                auction.item.current_price += 1
-                user.balance -= 1
+            if user.wallet.balance >= auction.current_price + PRICE_CHANGE:
+                auction.current_price += PRICE_CHANGE
                 user.save()
-                auction.item.save()
+                auction.save()
                 print(f"New bid: {
-                      auction.item.current_price} by {user.name}")
+                      auction.current_price} by {user.name}")
             else:
                 print("Insufficient funds")
         except (Auction.DoesNotExist, User.DoesNotExist) as e:
